+++ bootique-logback-sentry/src/main/java/io/bootique/logback/sentry/LogbackSentryModule.java
@@ -6,8 +6,6 @@ import com.google.inject.Provides;
 import com.google.inject.Singleton;
 import io.sentry.Sentry;
 import io.sentry.SentryClientFactory;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Loads configuration for {@link io.sentry.Sentry}.
@@ -16,7 +14,6 @@ import org.slf4j.LoggerFactory;
  * @since 0.16
  */
 public class LogbackSentryModule implements Module {
-    private final static Logger LOGGER = LoggerFactory.getLogger(LogbackSentryModule.class);
 
     @Override
     public void configure(Binder binder) {