+++ bootique-logback-sentry/src/main/java/io/bootique/logback/sentry/LogbackSentryModuleProvider.java
@@ -2,6 +2,10 @@ package io.bootique.logback.sentry;
 
 import com.google.inject.Module;
 import io.bootique.BQModuleProvider;
+import io.bootique.logback.LogbackModuleProvider;
+
+import java.util.Collection;
+import java.util.Collections;
 
 /**
  * Provider for {@link LogbackSentryModule}.
@@ -14,5 +18,10 @@ public class LogbackSentryModuleProvider implements BQModuleProvider {
     @Override
     public Module module() {
         return new LogbackSentryModule();
+    }
+
+    @Override
+    public Collection<BQModuleProvider> dependencies() {
+        return Collections.singletonList(new LogbackModuleProvider());
     }
 }